Restore missing graphics properties.
Also, increase debug.hwc.dynThreshold to 4.5

This value indicates the total pixel area across all layers can be
composited using copybit. This is as a factor of the framebuffer size.

The original value was calibrated for phones and tablets. Because
wearables have smaller framebuffers this ratio can be safely increased.

Change-Id: Ie13fe03352b8a06e88b2799e58a784ffff91c1ea