Disable backpressure propagation in SurfaceFlinger for 8x09 devices.
Because the msm8x09 hardware composer has a total latency of
up to 30ms (from 'handleMessageInvalidate' until the image appears
on the screen) backpressure propagation causes us to run at 30Hz.

This change allows us to run smoothly at 60Hz.

Also note that we should be able to revert this change in the event
that we switch to a native HWC2.0 implementation.

Fixes bug: 37941536

Change-Id: Ie045432e4de72ab59b206ae324e26d54795c9b8a